extracted errors and enumerations into separate files - fixed thread priority constants
