+++ mini_al/mini_al.go
@@ -45,236 +45,6 @@ import (
 	"unsafe"
 )
 
-// Return codes.
-const (
-	Success                        = 0
-	Error                          = -1
-	InvalidArgs                    = -2
-	InvalidOperation               = -3
-	OutOfMemory                    = -4
-	FormatNotSupported             = -5
-	NoBackend                      = -6
-	NoDevice                       = -7
-	APINotFound                    = -8
-	DeviceBusy                     = -9
-	DeviceNotInitialized           = -10
-	DeviceNotStarted               = -11
-	DeviceNotStopped               = -12
-	DeviceAlreadyStarted           = -13
-	DeviceAlreadyStarting          = -14
-	DeviceAlreadyStopped           = -15
-	DeviceAlreadyStopping          = -16
-	FailedToMapDeviceBuffer        = -17
-	FailedToUnmapDeviceBuffer      = -18
-	FailedToInitBackend            = -19
-	FailedToReadDataFromClient     = -20
-	FailedToReadDataFromDevice     = -21
-	FailedToSendDataToClient       = -22
-	FailedToSendDataToDevice       = -23
-	FailedToOpenBackendDevice      = -24
-	FailedToStartBackendDevice     = -25
-	FailedToStopBackendDevice      = -26
-	FailedToConfigureBackendDevice = -27
-	FailedToCreateMutex            = -28
-	FailedToCreateEvent            = -29
-	FailedToCreateThread           = -30
-	InvalidDeviceConfig            = -31
-	AccessDenied                   = -32
-	TooLarge                       = -33
-)
-
-// Errors.
-var (
-	errTag string = "mini_al"
-
-	ErrError                          = fmt.Errorf("%s: generic error", errTag)
-	ErrInvalidArgs                    = fmt.Errorf("%s: invalid args", errTag)
-	ErrInvalidOperation               = fmt.Errorf("%s: invalid operation", errTag)
-	ErrOutOfMemory                    = fmt.Errorf("%s: out of memory", errTag)
-	ErrFormatNotSupported             = fmt.Errorf("%s: format not supported", errTag)
-	ErrNoBackend                      = fmt.Errorf("%s: no backend", errTag)
-	ErrNoDevice                       = fmt.Errorf("%s: no device", errTag)
-	ErrAPINotFound                    = fmt.Errorf("%s: api not found", errTag)
-	ErrDeviceBusy                     = fmt.Errorf("%s: device busy", errTag)
-	ErrDeviceNotInitialized           = fmt.Errorf("%s: device busy", errTag)
-	ErrDeviceNotStarted               = fmt.Errorf("%s: device not started", errTag)
-	ErrDeviceNotStopped               = fmt.Errorf("%s: device not stopped", errTag)
-	ErrDeviceAlreadyStarted           = fmt.Errorf("%s: device already started", errTag)
-	ErrDeviceAlreadyStarting          = fmt.Errorf("%s: device already starting", errTag)
-	ErrDeviceAlreadyStopped           = fmt.Errorf("%s: device already stopped", errTag)
-	ErrDeviceAlreadyStopping          = fmt.Errorf("%s: device already stopping", errTag)
-	ErrFailedToMapDeviceBuffer        = fmt.Errorf("%s: failed to map device buffer", errTag)
-	ErrFailedToUnmapDeviceBuffer      = fmt.Errorf("%s: failed to unmap device buffer", errTag)
-	ErrFailedToInitBackend            = fmt.Errorf("%s: failed to init backend", errTag)
-	ErrFailedToReadDataFromClient     = fmt.Errorf("%s: failed to read data from client", errTag)
-	ErrFailedToReadDataFromDevice     = fmt.Errorf("%s: failed to read data from device", errTag)
-	ErrFailedToSendDataToClient       = fmt.Errorf("%s: failed to send data to client", errTag)
-	ErrFailedToSendDataToDevice       = fmt.Errorf("%s: failed to send data to device", errTag)
-	ErrFailedToOpenBackendDevice      = fmt.Errorf("%s: failed to open backend device", errTag)
-	ErrFailedToStartBackendDevice     = fmt.Errorf("%s: failed to start backend device", errTag)
-	ErrFailedToStopBackendDevice      = fmt.Errorf("%s: failed to stop backend device", errTag)
-	ErrFailedToConfigureBackendDevice = fmt.Errorf("%s: failed to configure backend device", errTag)
-	ErrFailedToCreateMutex            = fmt.Errorf("%s: failed to create mutex", errTag)
-	ErrFailedToCreateEvent            = fmt.Errorf("%s: failed to create event", errTag)
-	ErrFailedToCreateThread           = fmt.Errorf("%s: failed to create thread", errTag)
-	ErrInvalidDeviceConfig            = fmt.Errorf("%s: invalid device config", errTag)
-	ErrAccessDenied                   = fmt.Errorf("%s: access denied", errTag)
-	ErrTooLarge                       = fmt.Errorf("%s: too large", errTag)
-)
-
-// errorFromResult returns error for result code.
-func errorFromResult(r Result) error {
-	switch r {
-	case Success:
-		return nil
-	case Error:
-		return ErrError
-	case InvalidArgs:
-		return ErrInvalidArgs
-	case InvalidOperation:
-		return ErrInvalidOperation
-	case OutOfMemory:
-		return ErrOutOfMemory
-	case FormatNotSupported:
-		return ErrFormatNotSupported
-	case NoBackend:
-		return ErrNoBackend
-	case NoDevice:
-		return ErrNoDevice
-	case APINotFound:
-		return ErrAPINotFound
-	case DeviceBusy:
-		return ErrDeviceBusy
-	case DeviceNotInitialized:
-		return ErrDeviceNotInitialized
-	case DeviceNotStarted:
-		return ErrDeviceNotStarted
-	case DeviceNotStopped:
-		return ErrDeviceNotStopped
-	case DeviceAlreadyStarted:
-		return ErrDeviceAlreadyStarted
-	case DeviceAlreadyStarting:
-		return ErrDeviceAlreadyStarting
-	case DeviceAlreadyStopped:
-		return ErrDeviceAlreadyStopped
-	case DeviceAlreadyStopping:
-		return ErrDeviceAlreadyStopping
-	case FailedToMapDeviceBuffer:
-		return ErrFailedToMapDeviceBuffer
-	case FailedToUnmapDeviceBuffer:
-		return ErrFailedToUnmapDeviceBuffer
-	case FailedToInitBackend:
-		return ErrFailedToInitBackend
-	case FailedToReadDataFromClient:
-		return ErrFailedToReadDataFromClient
-	case FailedToReadDataFromDevice:
-		return ErrFailedToReadDataFromDevice
-	case FailedToSendDataToClient:
-		return ErrFailedToSendDataToClient
-	case FailedToSendDataToDevice:
-		return ErrFailedToSendDataToDevice
-	case FailedToOpenBackendDevice:
-		return ErrFailedToOpenBackendDevice
-	case FailedToStartBackendDevice:
-		return ErrFailedToStartBackendDevice
-	case FailedToStopBackendDevice:
-		return ErrFailedToStopBackendDevice
-	case FailedToConfigureBackendDevice:
-		return ErrFailedToConfigureBackendDevice
-	case FailedToCreateMutex:
-		return ErrFailedToCreateMutex
-	case FailedToCreateEvent:
-		return ErrFailedToCreateEvent
-	case FailedToCreateThread:
-		return ErrFailedToCreateThread
-	case InvalidDeviceConfig:
-		return ErrInvalidDeviceConfig
-	case AccessDenied:
-		return ErrAccessDenied
-	case TooLarge:
-		return ErrTooLarge
-	default:
-		return ErrError
-	}
-}
-
-// Backend type.
-type Backend uint32
-
-// Backend enumeration.
-const (
-	BackendNull Backend = iota
-	BackendWasapi
-	BackendDsound
-	BackendWinmm
-	BackendAlsa
-	BackendPulseAudio
-	BackendJack
-	BackendCoreAudio
-	BackendOss
-	BackendOpensl
-	BackendOpenal
-	BackendSdl
-)
-
-// DeviceType type.
-type DeviceType uint32
-
-// DeviceType enumeration.
-const (
-	Playback DeviceType = iota
-	Capture
-)
-
-// ShareMode type.
-type ShareMode uint32
-
-// ShareMode enumeration.
-const (
-	Shared ShareMode = iota
-	Exclusive
-)
-
-// PerformanceProfile type.
-type PerformanceProfile uint32
-
-// PerformanceProfile enumeration.
-const (
-	LowLatency PerformanceProfile = iota
-	Conservative
-)
-
-// FormatType type.
-type FormatType uint32
-
-// Format enumeration.
-const (
-	FormatUnknown FormatType = iota
-	FormatU8
-	FormatS16
-	FormatS24
-	FormatS32
-	FormatF32
-)
-
-// ThreadPriority type.
-type ThreadPriority int32
-
-// ThreadPriority enumeration.
-const (
-	Idle     ThreadPriority = -5
-	Lowest   ThreadPriority = -4
-	Low      ThreadPriority = -3
-	Normal   ThreadPriority = -2
-	High     ThreadPriority = -1
-	Highest  ThreadPriority = -0
-	Realtime ThreadPriority = -1
-	Default  ThreadPriority = -2
-)
-
-// Result type.
-type Result int32
-
 // Device type.
 type Device struct {
 	context *C.mal_context
